♻️ FQBWH9 task: split task backend branch snapshots
