+++ packages/agentplane/src/commands/shared/task-backend.ts
@@ -1,38 +1,26 @@
 import type { ResolvedProject } from "@agentplaneorg/core/project";
-import { readFile } from "node:fs/promises";
 import path from "node:path";
 import type { AgentplaneConfig } from "@agentplaneorg/core/config";
-import type { TaskRecord } from "@agentplaneorg/core/tasks";
-import {
-  resolveTaskDocUpdatedBy,
-  parseTaskReadme,
-  taskDocToSectionMap,
-} from "@agentplaneorg/core/tasks";
-import {
-  validateTaskReadmeFrontmatter,
-  withTaskReadmeFrontmatterDefaults,
-} from "@agentplaneorg/core/schemas";
+import { resolveTaskDocUpdatedBy, taskDocToSectionMap } from "@agentplaneorg/core/tasks";
 
 import type { ResolvedHarnessContract } from "../../runtime/harness/index.js";
 import { CliError } from "../../shared/errors.js";
 import { emitTraceEvent } from "../../shared/trace-events.js";
 import {
   loadTaskBackend,
-  taskRecordToData,
   type TaskBackendCapabilities,
   type TaskBackend,
   toTaskSummary,
   type TaskData,
   type TaskSummary,
 } from "../../backends/task-backend.js";
-import {
-  findWorktreeForBranch,
-  GitContext,
-  gitListTaskBranches,
-  gitShowFile,
-  parseTaskIdFromBranch,
-  toGitPath,
-} from "@agentplaneorg/core/git";
+import { GitContext } from "@agentplaneorg/core/git";
+import { loadTaskFromBranchSnapshot } from "./task-backend-branch-snapshot.js";
+
+export {
+  loadTaskFromBranchSnapshot,
+  resolveTaskBranchFromContext,
+} from "./task-backend-branch-snapshot.js";
 
 export type CommandMemo = {
   taskProjection?: Promise<TaskSummary[]>;
@@ -233,112 +221,6 @@ export async function loadTaskFromContext(opts: {
     exitCode: 4,
     code: "E_IO",
     message: `ENOENT: no such file or directory, open '${readmePath}'`,
-  });
-}
-
-export async function resolveTaskBranchFromContext(opts: {
-  ctx: CommandContext;
-  taskId: string;
-}): Promise<string | null> {
-  if (!backendSupportsTaskBranchSnapshots(opts.ctx)) {
-    return null;
-  }
-
-  const prefix = opts.ctx.config.branch.task_prefix;
-  const branches = await gitListTaskBranches(opts.ctx.resolvedProject.gitRoot, prefix);
-  const matches = branches.filter(
-    (branch) => parseTaskIdFromBranch(prefix, branch) === opts.taskId,
-  );
-  if (matches.length === 1) return matches[0] ?? null;
-  if (matches.length > 1) {
-    throw new CliError({
-      exitCode: 3,
-      code: "E_VALIDATION",
-      message: `Multiple task branches match ${opts.taskId}: ${matches.join(", ")}`,
-    });
-  }
-  return null;
-}
-
-export async function loadTaskFromBranchSnapshot(opts: {
-  ctx: CommandContext;
-  taskId: string;
-  readmePath: string;
-  branch?: string | null;
-}): Promise<TaskData | null> {
-  if (!backendSupportsTaskBranchSnapshots(opts.ctx)) {
-    return null;
-  }
-
-  const branch =
-    typeof opts.branch === "string" && opts.branch.trim().length > 0
-      ? opts.branch.trim()
-      : await resolveTaskBranchFromContext({ ctx: opts.ctx, taskId: opts.taskId });
-  if (!branch) return null;
-
-  const liveWorktreePath = await findWorktreeForBranch(opts.ctx.resolvedProject.gitRoot, branch);
-  if (liveWorktreePath) {
-    const liveReadmePath = path.join(
-      liveWorktreePath,
-      path.relative(opts.ctx.resolvedProject.gitRoot, opts.readmePath),
-    );
-    try {
-      const liveText = await readFile(liveReadmePath, "utf8");
-      const parsed = parseTaskReadme(liveText);
-      const frontmatter = validateTaskReadmeFrontmatter(
-        withTaskReadmeFrontmatterDefaults({
-          ...parsed.frontmatter,
-          id:
-            typeof parsed.frontmatter.id === "string" && parsed.frontmatter.id.trim().length > 0
-              ? parsed.frontmatter.id
-              : opts.taskId,
-        }),
-      );
-      return taskRecordToData({
-        id: opts.taskId,
-        frontmatter: frontmatter as unknown as TaskRecord["frontmatter"],
-        body: parsed.body,
-        readmePath: opts.readmePath,
-      });
-    } catch {
-      // Fall back to the committed branch snapshot when the live worktree lacks the README.
-    }
-  }
-
-  const relReadmePath = toGitPath(path.relative(opts.ctx.resolvedProject.gitRoot, opts.readmePath));
-
-  const refsToTry = [branch, branch.startsWith("origin/") ? null : `origin/${branch}`].filter(
-    (ref): ref is string => Boolean(ref && ref.trim().length > 0),
-  );
-
-  let text = "";
-  let loaded = false;
-  for (const ref of refsToTry) {
-    try {
-      text = await gitShowFile(opts.ctx.resolvedProject.gitRoot, ref, relReadmePath);
-      loaded = true;
-      break;
-    } catch {
-      // Try the next candidate ref.
-    }
-  }
-  if (!loaded) return null;
-
-  const parsed = parseTaskReadme(text);
-  const frontmatter = validateTaskReadmeFrontmatter(
-    withTaskReadmeFrontmatterDefaults({
-      ...parsed.frontmatter,
-      id:
-        typeof parsed.frontmatter.id === "string" && parsed.frontmatter.id.trim().length > 0
-          ? parsed.frontmatter.id
-          : opts.taskId,
-    }),
-  );
-  return taskRecordToData({
-    id: opts.taskId,
-    frontmatter: frontmatter as unknown as TaskRecord["frontmatter"],
-    body: parsed.body,
-    readmePath: opts.readmePath,
   });
 }
